Update index.html
Removed the default 0's from the html text, as the jquery code waits for the page to finish loading 100% before executing code. This ended up being about 1 second on my end, and with loading code, you'd be seeing 0's when you may have already played and know that you have more than 0. So, it's technically true that we are loading, but not during that 1 initial second, but nobody needs to know that ;)