Update main.js
restructured code, putting functions inside the main document.load() function. i did this initially to see if variables declared within those functions would be considered local since they, along with all the other code, are inside the document.load() function. it doesn't work. that's why i've set up the very first line as some global variables. it may be messy as it grows, but the alternative is to use 'window.whatevervariable' on every single variable, which just sucks.

added in debugging options. right now there's the DEBUG variable which can be manually set to 1 to turn on debugging options or 0 to turn them off. there's an 'if (DEBUGGING ===1) {}' code block that will enable things when we need them. currently there is a button that can delete the localStorage for the game and start over. at some point perhaps we'll decide to make it a game feature, along with a "are you very very sure you want to do this, because if you are, you are not allowed to yell at us for having all your progress wiped" kinda disclaimer :P