Update main.js
Changed the timing of the main loop - it now refreshes every 10ms.
This will allow changes to appear immediately - if you recall, with the previous loop there was a delay in displaying the house button once the appropriate resources were gathered. This now shows up immediately.

We still have things that need to be updated strictly once per second, and so these items are now in their own section in the loop, being controlled by a 'millisecondsElapsed' variable. Each time we go through the loop, we increment this variable by 10 (the same amount of time we're using in the loop - 10ms). Once the variable's value is at 1000, 1 second has passed, and we can execute the appropriate code.